I created subfolder robust-ts in folders [branches/robust-ts-<...>/]pkg (where <...> stands for bs or do, i.e. robust-ts is a new subfolder of pkg and the respective versions in branches robust-ts-bs, robust-ts-do.)
This new subfolder will ease work with R CMD check, R CMD build, R CMD INSTALL 
on your working copy of the repository afterwords:

You may simply go to the corresponding [branches/robust-ts-<...>]/pkg
folder, and in this folder call
R CMD check robust-ts    (and correspondingly for build and INSTALL)

To this end, I moved all contents of the respective [branches/robust-ts-<...>/]pkg 
folder to the respective [branches/robust-ts-<...>/]pkg/robKalman folder, but
but not the svn system folders (which after one commitment of Bernhard is not yet 
informative, anyway). 
So in particular the contents of of the respective [branches/robust-ts-<...>/]pkg/R 
are deleted and are now in folder [branches/robust-ts-<...>/]pkg/robust-ts/R

Subsequent work should simply be done in the respective 
        [branches/robust-ts-<...>/]pkg/robust-ts 
folder.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/robust-ts@17 edb9625f-4e0d-4859-8d74-9fd3b1da38cb